requirements.txt
upstash-vector
groq
python-dotenv
requests
psycopg2-binary
pgvector
tenacity